Clean up Gradle usage
* Remove buildSrc
* Move all plugins to plugins in version catalog
