--- buildSrc/settings.gradle.kts
@@ -1,23 +1,0 @@
-/**
- * Copyright © 2025 Apple Inc. and the Pkl project authors. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-dependencyResolutionManagement {
-  // use same version catalog as main build
-  versionCatalogs { register("libs") { from(files("../gradle/libs.versions.toml")) } }
-
-  @Suppress("UnstableApiUsage")
-  repositories {
-    mavenCentral()
-    gradlePluginPortal()
-  }
-}
