--- buildSrc/src/main/kotlin/PklFormatterSpotless.kt
@@ -1,43 +1,0 @@
-/**
- * Copyright © 2025 Apple Inc. and the Pkl project authors. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-import com.diffplug.spotless.FormatterFunc
-import com.diffplug.spotless.FormatterStep
-import java.io.Serial
-import java.io.Serializable
-import org.pkl.formatter.Formatter
-import org.pkl.formatter.GrammarVersion
-
-class PklFormatterStep() : Serializable {
-  companion object {
-    @Serial private const val serialVersionUID: Long = 1L
-  }
-
-  fun create(): FormatterStep {
-    return FormatterStep.createLazy(
-      "pkl",
-      { PklFormatterStep() },
-      { PklFormatterFunc() },
-    )
-  }
-}
-
-class PklFormatterFunc() : FormatterFunc, Serializable {
-  companion object {
-    @Serial private const val serialVersionUID: Long = 1L
-  }
-
-  override fun apply(input: String): String {
-    return Formatter().format(input, GrammarVersion.V1)
-  }
-}
